// src/lib/pa-logos.generated.ts
export const PA_LOGO_FILES = {
  abby: "abby.png",
  accenture: "accenture.png",
  "ademico-software": "ademico-software.png",
  "agena-3000": "agena-3000.png",
  agicap: "agicap.png",
  "ags-records-management": "ags-records-management.png",
  arratech: "arratech.png",
  arteva: "arteva.png",
  aruba: "aruba.png",
  atgp: "atgp.png",
  avalara: "avalara.png",
  axelor: "axelor.png",
  axonaut: "axonaut.png",
  "axway-software": "axway-software.png",
  azopio: "azopio.jpg",
  b2brouter: "b2brouter.png",
  "b4value-net": "b4value-net.png",
  basware: "basware.png",
  "bc-solutions": "bc-solutions.png",
  billit: "billit.png",
  "cbs-corporate-business-solutions": "cbs-corporate-business-solutions.png",
  cecurity: "cecurity.png",
  cegedim: "cegedim.png",
  "cegi-alfa": "cegi-alfa.png",
  cegid: "cegid.png",
  chaintrust: "chaintrust.png",
  cleartax: "cleartax.png",
  comarch: "comarch.png",
  darva: "darva.png",
  dext: "dext.png",
  digipharmacie: "digipharmacie.png",
  "digital-technologies": "digital-technologies.png",
  docoon: "docoon.png",
  "docoon-immo-freedz": "docoon-immo-freedz.png",
  docprocess: "docprocess.png",
  docuware: "docuware.png",
  dokapi: "dokapi.png",
  dougs: "dougs.png",
  doxallia: "doxallia.png",
  ebp: "ebp.jpg",
  ecosio: "ecosio.png",
  "ecosio-intercom-vertex": "ecosio-intercom-vertex.png",
  "edicom-france": "edicom-france.png",
  "edicom-group": "edicom-group.png",
  "edics-france": "edics-france.png",
  "edieyes-vision-care": "edieyes-vision-care.png",
  edt: "edt.png",
  eezi: "eezi.png",
  effinum: "effinum.png",
  enerj: "enerj.jpg",
  entropics: "entropics.png",
  esalink: "esalink.png",
  esi: "esi.png",
  esker: "esker.png",
  "euro-information": "euro-information.png",
  facnote: "facnote.png",
  "fiducial-cloud": "fiducial-cloud.png",
  flowie: "flowie.png",
  fonoa: "fonoa.png",
  freebe: "freebe.jpg",
  fulll: "fulll.png",
  "generix-group": "generix-group.png",
  gestav: "gestav.png",
  "groupe-sigma": "groupe-sigma.png",
  gurusoft: "gurusoft.png",
  henrri: "henrri.png",
  iaf: "iaf.png",
  "icd-international": "icd-international.png",
  iedi: "iedi.png",
  "iga-assurance": "iga-assurance.svg",
  indicom: "indicom.png",
  indy: "indy.png",
  infologic: "infologic.png",
  invopop: "invopop.png",
  iopole: "iopole.png",
  ipaidthat: "ipaidthat.svg",
  itesoft: "itesoft.png",
  jefacture: "jefacture.png",
  klekoon: "klekoon.png",
  "kolecto-pa": "kolecto-pa.png",
  "kolecto-pdp": "kolecto-pdp.png",
  "le-village-connecte": "le-village-connecte.png",
  "legalinvoice-tinexta-infocert": "legalinvoice-tinexta-infocert.png",
  logilec: "logilec.png",
  lucca: "lucca.png",
  "macompta-fr": "macompta-fr.png",
  marosa: "marosa.png",
  medius: "medius.png",
  melasoft: "melasoft.png",
  "my-fiteco": "my-fiteco.png",
  "my-unisoft": "my-unisoft.png",
  mykinexo: "mykinexo.png",
  mysupply: "mysupply.png",
  "n2f-pdp": "n2f-pdp.png",
  neotimo: "neotimo.png",
  newblack: "newblack.png",
  "ntt-data-business-solutions": "ntt-data-business-solutions.png",
  numeria: "numeria.png",
  odoo: "odoo.png",
  "one-up": "one-up.png",
  "open-bee": "open-bee.png",
  opentext: "opentext.png",
  opuscapita: "opuscapita.png",
  pagero: "pagero.png",
  paragon: "paragon.png",
  payflows: "payflows.png",
  pennylane: "pennylane.png",
  "pitney-bowes": "pitney-bowes.png",
  qonto: "qonto.png",
  "quadient-serensia": "quadient-serensia.png",
  qweeby: "qweeby.png",
  ratioo: "ratioo.png",
  sage: "sage.png",
  sap: "sap.png",
  scribee: "scribee.png",
  seeburger: "seeburger.png",
  sellsy: "sellsy.png",
  septeo: "septeo.png",
  seqino: "seqino.jpg",
  seres: "seres.png",
  shine: "shine.png",
  sni: "sni.png",
  solo: "solo.png",
  sovos: "sovos.png",
  spendesk: "spendesk.png",
  "sps-commerce": "sps-commerce.png",
  srci: "srci.png",
  storecove: "storecove.png",
  "super-pdp": "super-pdp.png",
  symtrax: "symtrax.png",
  taxera: "taxera.png",
  tenor: "tenor.png",
  tesisquare: "tesisquare.png",
  tessi: "tessi.png",
  tiime: "tiime.png",
  "tradeshift-babelway": "tradeshift-babelway.png",
  transalis: "transalis.png",
  treso2: "treso2.png",
  "tungsten-automation-france": "tungsten-automation-france.jpg",
  "tx2-concept": "tx2-concept.jpg",
  ventya: "ventya.png",
  veryswing: "veryswing.png",
  vosfactures: "vosfactures.png",
  "voxel-amadeus": "voxel-amadeus.jpg",
  wakastellar: "wakastellar.png",
  weproc: "weproc.png",
  "wisetech-global": "wisetech-global.png",
  xelya: "xelya.png",
  yooz: "yooz.png",
} as const;